okay guys I was bored and didnt have anything to do
so the game is not over when the sheep are dead....

get this the wolfs start playing tag with each other as soon as there 
are no more sheep.

oh and if you want to revive the sheeps just hit the hunt button and 
it revives the sheeps 

since they are still moving in the background they just dont get
displayed.