okay wolfs and sheeps now move a bit better.
also they play tag a bit better.
of course after the sheep all die....

if you hit the hunt button it rivives one of the sheeps 
and the wolfs stop playing tag and get into attack mode......

sheep now start moving when the see a wolf near them and but they are a
bit slower still messing with that.