just adding some of the changes
commands are still start of with 

"i am your father"

root is able to run these commands

move.3,4=4,5
pretty much moves what ever is there too the other side of the =

add.sheep=3,4

adds a sheep trying to get the add wolf to work right now.

start.game
or start whatever makes the game start.

and that is all